fix static-analysis failure
