+++ tests/vxlan_test.py
@@ -406,59 +406,3 @@ Error: 'vxlan_name' length should not exceed 15 characters
     def teardown_class(cls):
         os.environ['UTILITIES_UNIT_TESTING'] = "0"
         print("TEARDOWN")
-
-
-class TestVnet(object):
-    @classmethod
-    def setup_class(cls):
-        os.environ['UTILITIES_UNIT_TESTING'] = "1"
-        print("SETUP")
-
-    def test_show_vnet_brief(self):
-        from .mock_tables import dbconnector
-        jsonfile_config = os.path.join(mock_db_path, "config_db")
-        dbconnector.dedicated_dbs['CONFIG_DB'] = jsonfile_config
-        runner = CliRunner()
-
-        result = runner.invoke(show.cli.commands["vnet"].commands["brief"], [])
-        print(result.output)
-        assert result.exit_code == 0
-        assert "Vnet_2000" in result.output
-        assert "1234-56-7890-1234" in result.output
-        assert "tunnel1" in result.output
-
-    def test_show_vnet_name(self):
-        from .mock_tables import dbconnector
-        jsonfile_config = os.path.join(mock_db_path, "config_db")
-        dbconnector.dedicated_dbs['CONFIG_DB'] = jsonfile_config
-        runner = CliRunner()
-
-        result = runner.invoke(show.cli.commands["vnet"].commands["name"], ["Vnet_2000"])
-        print(result.output)
-        assert result.exit_code == 0
-        assert "Vnet_2000" in result.output
-        assert "1234-56-7890-1234" in result.output
-        assert "Ethernet4" in result.output
-        assert "Vlan100" in result.output
-
-    def test_show_vnet_guid(self):
-        from .mock_tables import dbconnector
-        jsonfile_config = os.path.join(mock_db_path, "config_db")
-        dbconnector.dedicated_dbs['CONFIG_DB'] = jsonfile_config
-        runner = CliRunner()
-
-        result = runner.invoke(show.cli.commands["vnet"].commands["guid"], ["1234-56-7890-1234"])
-        print(result.output)
-        assert result.exit_code == 0
-        assert "Vnet_2000" in result.output
-        assert "1234-56-7890-1234" in result.output
-        assert "tunnel1" in result.output
-        assert "Ethernet4" in result.output
-        assert "Vlan100" in result.output
-
-    @classmethod
-    def teardown_class(cls):
-        os.environ['UTILITIES_UNIT_TESTING'] = "0"
-        print("TEARDOWN")
-        from .mock_tables import dbconnector
-        dbconnector.dedicated_dbs = {}
